Plant_Disease_model/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
python-multipart==0.0.6
pydantic==1.10.7
